refactor: commit history chart
- reuse filter logic in tree/hardware details and commit chart
- remove builds, tests, boots counts CTE from query
- change the query to get the earliest commits
- compute count in python

Part of #640
