+++ backend/kernelCI_app/views/treeCommitsHistory.py
@@ -1,13 +1,16 @@
+from kernelCI_app.utils import (
+    getErrorResponseBody
+)
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from django.http import JsonResponse, HttpResponseBadRequest
 from django.db import connection
-from kernelCI_app.utils import (
-    NULL_STRINGS,
+from typing import Dict, List
+from kernelCI_app.helpers.filters import (
+    UNKNOWN_STRING,
     FilterParams,
-    InvalidComparisonOP,
-    getErrorResponseBody,
+    InvalidComparisonOP
 )
 
 
@@ -182,165 +185,67 @@ class TreeCommitsHistory(APIView):
             "git_branch_param": git_branch_param,
         }
 
-        build_counts_where, boot_counts_where, test_counts_where = self.__get_filters(filter_params)
-
-        query = f"""
-        WITH
-        relevant_checkouts AS (
+        query = """
+        WITH earliest_commits AS (
             SELECT
-                id,
+                 id,
                 git_commit_hash,
                 git_commit_name,
                 git_repository_branch,
                 git_repository_url,
                 origin,
-                start_time
-            FROM
-                checkouts
+                start_time,
+                time_order
+            FROM (
+                SELECT
+                    id,
+                    git_commit_hash,
+                    git_commit_name,
+                    git_repository_branch,
+                    git_repository_url,
+                    origin,
+                    start_time,
+                    ROW_NUMBER() OVER (
+                        PARTITION BY git_repository_url, git_repository_branch, origin, git_commit_hash
+                        ORDER BY start_time DESC
+                    ) AS time_order
+                FROM   checkouts
+                WHERE  git_repository_branch = %(git_branch_param)s
+                    AND git_repository_url = %(git_url_param)s
+                    AND origin = %(origin_param)s
+                    AND git_commit_hash IS NOT NULL
+                    AND start_time <= (SELECT Max(start_time) AS head_start_time
+                                        FROM   checkouts
+                                        WHERE  git_commit_hash = %(commit_hash)s
+                                                AND origin = %(origin_param)s
+                                                AND git_repository_url = %(git_url_param)s)
+                ORDER  BY start_time DESC) AS checkouts_time_order
             WHERE
-                git_repository_branch = %(git_branch_param)s
-                AND git_repository_url = %(git_url_param)s
-                AND origin = %(origin_param)s
-                AND git_commit_hash IS NOT NULL
-                AND start_time <= (
-                    SELECT MAX(start_time) as head_start_time
-                    FROM checkouts
-                    WHERE git_commit_hash = %(commit_hash)s
-                )
-            ORDER BY
-                start_time DESC
-        ),
-        earliest_commits AS (
-            SELECT DISTINCT ON (git_commit_hash)
-                git_commit_hash,
-                git_commit_name,
-                MAX(start_time) AS earliest_start_time
-            FROM
-                relevant_checkouts
-            GROUP BY
-                git_commit_hash,
-                git_commit_name
-            ORDER BY
-                git_commit_hash
-        ),
-        build_counts AS (
-            SELECT
-                c.git_commit_hash,
-                SUM(CASE WHEN b.valid = true THEN 1 ELSE 0 END) AS valid_builds,
-                SUM(CASE WHEN b.valid = false THEN 1 ELSE 0 END) AS invalid_builds,
-                SUM(CASE WHEN b.valid IS NULL THEN 1 ELSE 0 END) AS null_builds
-            FROM
-                relevant_checkouts AS c
-            INNER JOIN
-                builds AS b
-                ON c.id = b.checkout_id
-            WHERE
-                {build_counts_where}
-            GROUP BY
-                c.git_commit_hash
-        ),
-        boots_counts AS (
-            SELECT
-                c.git_commit_hash,
-                SUM(CASE WHEN t.status = 'FAIL'
-                    AND (t.path = 'boot' OR t.path LIKE 'boot.%%') THEN 1 ELSE 0 END)
-                    AS boots_fail_count,
-                SUM(CASE WHEN t.status = 'ERROR'
-                    AND (t.path = 'boot' OR t.path LIKE 'boot.%%') THEN 1 ELSE 0 END)
-                    AS boots_error_count,
-                SUM(CASE WHEN t.status = 'MISS'
-                    AND (t.path = 'boot' OR t.path LIKE 'boot.%%') THEN 1 ELSE 0 END)
-                    AS boots_miss_count,
-                SUM(CASE WHEN t.status = 'PASS'
-                    AND (t.path = 'boot' OR t.path LIKE 'boot.%%') THEN 1 ELSE 0 END)
-                    AS boots_pass_count,
-                SUM(CASE WHEN t.status = 'DONE'
-                    AND (t.path = 'boot' OR t.path LIKE 'boot.%%') THEN 1 ELSE 0 END)
-                    AS boots_done_count,
-                SUM(CASE WHEN t.status = 'SKIP'
-                    AND (t.path = 'boot' OR t.path LIKE 'boot.%%') THEN 1 ELSE 0 END)
-                    AS boots_skip_count,
-                SUM(CASE WHEN t.status IS NULL
-                    AND (t.path = 'boot' OR t.path LIKE 'boot.%%') THEN 1 ELSE 0 END)
-                    AS boots_null_count
-            FROM
-                relevant_checkouts AS c
-            INNER JOIN
-                builds AS b
-                ON c.id = b.checkout_id
-            LEFT JOIN
-                tests AS t
-                ON b.id = t.build_id
-            WHERE
-                {boot_counts_where}
-            GROUP BY
-                c.git_commit_hash
-        ),
-        test_counts AS (
-            SELECT
-                c.git_commit_hash,
-                SUM(CASE WHEN t.status = 'FAIL' AND (t.path <> 'boot' AND t.path NOT LIKE 'boot.%%')
-                    THEN 1 ELSE 0 END) AS non_boots_fail_count,
-                SUM(CASE WHEN t.status = 'ERROR' AND (t.path <> 'boot' AND t.path NOT LIKE 'boot.%%')
-                    THEN 1 ELSE 0 END) AS non_boots_error_count,
-                SUM(CASE WHEN t.status = 'MISS' AND (t.path <> 'boot' AND t.path NOT LIKE 'boot.%%')
-                    THEN 1 ELSE 0 END) AS non_boots_miss_count,
-                SUM(CASE WHEN t.status = 'PASS' AND (t.path <> 'boot' AND t.path NOT LIKE 'boot.%%')
-                    THEN 1 ELSE 0 END) AS non_boots_pass_count,
-                SUM(CASE WHEN t.status = 'DONE' AND (t.path <> 'boot' AND t.path NOT LIKE 'boot.%%')
-                    THEN 1 ELSE 0 END) AS non_boots_done_count,
-                SUM(CASE WHEN t.status = 'SKIP' AND (t.path <> 'boot' AND t.path NOT LIKE 'boot.%%')
-                    THEN 1 ELSE 0 END) AS non_boots_skip_count,
-                SUM(CASE WHEN t.status IS NULL AND (t.path <> 'boot' AND t.path NOT LIKE 'boot.%%')
-                    THEN 1 ELSE 0 END) AS non_boots_null_count
-            FROM
-                relevant_checkouts AS c
-            INNER JOIN
-                builds AS b
-                ON c.id = b.checkout_id
-            LEFT JOIN
-                tests AS t
-                ON b.id = t.build_id
-            WHERE
-                {test_counts_where}
-            GROUP BY
-                c.git_commit_hash
+                time_order = 1
+            LIMIT 6
         )
         SELECT
-            ec.git_commit_hash,
-            ec.git_commit_name,
-            ec.earliest_start_time,
-            bc.valid_builds,
-            bc.invalid_builds,
-            bc.null_builds,
-            boc.boots_fail_count,
-            boc.boots_error_count,
-            boc.boots_miss_count,
-            boc.boots_pass_count,
-            boc.boots_done_count,
-            boc.boots_skip_count,
-            boc.boots_null_count,
-            tc.non_boots_fail_count,
-            tc.non_boots_error_count,
-            tc.non_boots_miss_count,
-            tc.non_boots_pass_count,
-            tc.non_boots_done_count,
-            tc.non_boots_skip_count,
-            tc.non_boots_null_count
-        FROM
-            earliest_commits AS ec
-        LEFT JOIN
-            build_counts AS bc
-            ON ec.git_commit_hash = bc.git_commit_hash
-        LEFT JOIN
-            boots_counts AS boc
-            ON ec.git_commit_hash = boc.git_commit_hash
-        LEFT JOIN
-            test_counts AS tc
-            ON ec.git_commit_hash = tc.git_commit_hash
-        ORDER BY
-            ec.earliest_start_time DESC
-        LIMIT 6;
+            c.git_commit_hash,
+            c.git_commit_name,
+            c.start_time,
+            b.duration,
+            b.architecture,
+            b.compiler,
+            b.config_name,
+            b.valid,
+            t.path,
+            t.status,
+            t.duration,
+            t.environment_compatible,
+            b.id AS build_id,
+            t.id AS test_id
+        FROM earliest_commits AS c
+        INNER JOIN builds AS b
+        ON
+            c.id = b.checkout_id
+        LEFT JOIN tests AS t
+        ON
+            t.build_id = b.id
         """
 
         # Execute the query
@@ -351,37 +256,38 @@ class TreeCommitsHistory(APIView):
             )
             rows = cursor.fetchall()
 
+        self._process_rows(rows)
         # Format the results as JSON
-        results = [
-            {
-                "git_commit_hash": row[0],
-                "git_commit_name": row[1],
-                "earliest_start_time": row[2],
+        results = []
+
+        for key, value in self.commit_hashes.items():
+            results.append({
+                "git_commit_hash": key,
+                "git_commit_name": value['commit_name'],
+                "earliest_start_time": value['earliest_start_time'],
                 "builds": {
-                    "valid_builds": row[3] or 0,
-                    "invalid_builds": row[4] or 0,
-                    "null_builds": row[5] or 0,
+                    "valid_builds": value['builds_count']['true'],
+                    "invalid_builds": value['builds_count']['false'],
+                    "null_builds": value['builds_count']['null'],
                 },
                 "boots_tests": {
-                    "fail_count": row[6] or 0,
-                    "error_count": row[7] or 0,
-                    "miss_count": row[8] or 0,
-                    "pass_count": row[9] or 0,
-                    "done_count": row[10] or 0,
-                    "skip_count": row[11] or 0,
-                    "null_count": row[12] or 0,
+                    "fail_count": value['boots_count']['fail'],
+                    "error_count": value['boots_count']['error'],
+                    "miss_count": value['boots_count']['miss'],
+                    "pass_count": value['boots_count']['pass'],
+                    "done_count": value['boots_count']['done'],
+                    "skip_count": value['boots_count']['skip'],
+                    "null_count": value['boots_count']['null'],
                 },
                 "non_boots_tests": {
-                    "fail_count": row[13] or 0,
-                    "error_count": row[14] or 0,
-                    "miss_count": row[15] or 0,
-                    "pass_count": row[16] or 0,
-                    "done_count": row[17] or 0,
-                    "skip_count": row[18] or 0,
-                    "null_count": row[19] or 0,
+                    "fail_count": value['tests_count']['fail'],
+                    "error_count": value['tests_count']['error'],
+                    "miss_count": value['tests_count']['miss'],
+                    "pass_count": value['tests_count']['pass'],
+                    "done_count": value['tests_count']['done'],
+                    "skip_count": value['tests_count']['skip'],
+                    "null_count": value['tests_count']['null'],
                 },
-            }
-            for row in rows
-        ]
+            })
 
         return JsonResponse(results, safe=False)
